+++ packages/terra-table/CHANGELOG.md
@@ -3,11 +3,26 @@
 ## Unreleased
 
 * Added
+  * Added row selection mode to the table component.
+  * Added ability for consumers to specify a custom screen reader message for masked cells.
+
+* Fixed
+  * Fixed the column header background color for the Orion Fusion theme.
+  
+## 5.1.1-alpha.0 - (October 25, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 5.1.0-alpha.0 - (October 20, 2023)
+
+* Added
   * Added the ability to toggle zebra striping for table rows.
   * Added sections support to the table component.
   
 * Changed
   * Updated the table component so that the cell dive-in logic would not execute when not in the grid context.
+  * Updated the table component to allow consumers to control the visibility of column headers in the table.
   * Modified the table component so that it can receive focus when scrollable.
 
 ## 5.0.0-alpha.0 - (October 17, 2023)