+++ packages/terra-tabs/CHANGELOG.md
@@ -2,6 +2,15 @@
 
 ## Unreleased
 
+## 7.11.1 - (October 25, 2023)
+
+* Fixed
+  * Fixed the dropdown options are being cut off on smaller viewports.
+  * Fixed keyboard navigation after mouse click for draggable tabs.
+  * Fixed screen reader response when active tabs are closed.
+
+## 7.11.0 - (October 20, 2023)
+
 * Added
   * Added Vertical orientation for `terra-tabs`.